[clean] ファイルの階層構造変更
